refactor(link): colocate resolver context APIs
