+++ extensions/link/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./Link";
-export type { LinkUrlResolver } from "./LinkUrlResolverContext";
-export { LinkUrlResolverContext } from "./LinkUrlResolverContext";
+export * from "./LinkUrlResolverContext";
